Allow configuring custom slack
Resolves #20

This is to allow people to configure custom slack - we previously
defaulted to 10, and only allowed to disable it.

I'm changing the config structure a bit for it to describe the intention
better - it was previously storing `MaxSlack` as `Duration`,
but it should be an `int`.
Reasoning: we configure the limiter in terms of "operations per unit
of time" - like, 10 per second.
Similarly, with slack we should be saying: "10, with slack of 2, per second".
Only in the internals of the limiter the 10-per-second get converted
to Duration of 100ms.

Also had to modify how "maxSlack" is calculated.
Previously we just hardcoded it as "slack per second",
which didn't play nicely with "Per" option.
(so, technically, before this diff if you configured "10 per minute",
by default you'd get "10 per minute, with a slack of 10 per
second".

Extending existing slack test for all these.
The test is a bit weird, so tried to add a description.